fix: always cast emitted modelValue to number
