+++ src/components/NumberInput.vue
@@ -1,14 +1,48 @@
 <script lang="ts">
-import { defineComponent, ref } from 'vue'
-import AInput from './internal/Input.vue'
+import { defineComponent, PropType, ref, computed } from 'vue'
 import StepDownArrow from './internal/StepDownArrow.vue'
 import StepUpArrow from './internal/StepUpArrow.vue'
 
 export default defineComponent({
   name: 'ANumberInput',
-  components: { AInput, StepUpArrow, StepDownArrow },
+  components: { StepUpArrow, StepDownArrow },
   inheritAttrs: false,
   props: {
+    /**
+     * the prop modelValue is required to use v-model
+     */
+    modelValue: {
+      type: Number,
+      default: 0
+    },
+    /**
+     * the size of the input field
+     */
+    size: {
+      type: String as PropType<'sm' | 'md'>,
+      default: 'sm'
+    },
+    /**
+     * whether the input value is valid
+     */
+    invalid: {
+      type: Boolean,
+      default: false
+    },
+    /**
+     * how the input is displayed when not focused
+     */
+    variant: {
+      type: String as PropType<'subtle' | 'default'>,
+      default: 'default'
+    },
+    /**
+     * placeholder string
+     */
+    placeholder: {
+      type: String,
+      default: ''
+    },
     /**
      * works the same as the HTML step attribute, default `1`
      */
@@ -30,34 +64,63 @@ export default defineComponent({
     }
   },
   emits: ['update:model-value', 'update:by-step'],
-  setup(props, { expose, emit }) {
-    const inputRef = ref()
+  setup(props, { emit, expose, attrs }) {
+    const inputRef = ref<HTMLInputElement | null>(null)
 
-    const emitUpdateValue = (newValue: number) => {
-      emit('update:model-value', newValue)
+    const updateValue = (event: Event) => {
+      const target = event.target as HTMLInputElement
+      emit('update:model-value', target.valueAsNumber)
     }
+
+    const selectValue = () => {
+      if (props.size === 'sm' && inputRef.value) {
+        inputRef.value.select()
+      }
+    }
+
+    const placeholderText = computed(() => {
+      if (!props.placeholder && props.variant === 'subtle') {
+        if (attrs.type) {
+          return `Add ${attrs.type} value`
+        }
+        return 'Add value'
+      }
+      return props.placeholder
+    })
+
+    const variantClasses = computed(() => {
+      // invalid (error) state
+      if (props.invalid && !attrs.disabled) {
+        return 'a-text-input--invalid'
+      }
+      // subtle display variant
+      if (props.variant === 'subtle') {
+        return 'a-text-input--subtle'
+      }
+      // default display variant
+      return 'a-text-input--default'
+    })
 
     const emitUpdateByStep = (newValue: number) => {
       emit('update:by-step', newValue)
     }
 
     const useStep = (event: MouseEvent | KeyboardEvent, operation: 'up' | 'down') => {
-      const inputElement = inputRef.value?.el
-      if (inputElement) {
+      if (inputRef.value) {
         let step = props.step
         if (event.shiftKey && event.altKey && props.smallStep) {
           step = props.smallStep
         } else if (event.shiftKey && props.bigStep) {
           step = props.bigStep
         }
-        inputElement.step = step
+        inputRef.value.step = step.toString()
         if (operation === 'up') {
-          inputElement.stepUp()
+          inputRef.value.stepUp()
         } else {
-          inputElement.stepDown()
+          inputRef.value.stepDown()
         }
-        emitUpdateValue(inputElement.valueAsNumber)
-        emitUpdateByStep(inputElement.valueAsNumber)
+        emit('update:model-value', inputRef.value.valueAsNumber)
+        emitUpdateByStep(inputRef.value.valueAsNumber)
       }
     }
 
@@ -70,26 +133,40 @@ export default defineComponent({
     }
 
     expose({
-      inputRef: inputRef.value?.inputRef,
-      el: inputRef.value?.inputRef,
-      $el: inputRef.value?.inputRef
+      inputRef,
+      el: inputRef,
+      $el: inputRef
     })
 
-    return { stepUp, stepDown, inputRef, emitUpdateValue }
+    return {
+      stepUp,
+      stepDown,
+      inputRef,
+      updateValue,
+      selectValue,
+      placeholderText,
+      variantClasses
+    }
   }
 })
 </script>
+
 <template>
   <div class="group relative">
-    <AInput
-      v-bind="$attrs"
+    <input
       ref="inputRef"
-      :step="step"
       type="number"
-      class="a-number-input group-hover:a-text-input-hover"
+      class="a-text-input a-number-input group-hover:a-text-input-hover"
+      :class="[variantClasses, size === 'md' ? 'a-text-input--md' : 'a-text-input--sm']"
+      :value="modelValue"
+      :placeholder="placeholderText"
+      :aria-invalid="invalid"
+      :step="step"
+      v-bind="$attrs"
+      @input="updateValue"
+      @click="selectValue"
       @keydown.up.prevent="stepUp"
-      @keydown.down.prevent="stepDown"
-      @update:model-value="emitUpdateValue" />
+      @keydown.down.prevent="stepDown" />
     <div
       class="invisible absolute bottom-0 right-0 top-0 flex flex-col justify-center text-warsaw"
       :class="!$attrs.disabled && 'group-focus-within:visible group-hover:visible'">
