+++ src/stories/NumberInput.mdx
@@ -1,5 +1,5 @@
 import { Meta, ArgTypes, Canvas } from '@storybook/addon-docs'
-import AInput from '../components/internal/Input.vue'
+import ANumberInput from '../components/NumberInput.vue'
 
 import * as NumberInputStories from './NumberInput.stories'
 
@@ -14,8 +14,7 @@ import * as NumberInputStories from './NumberInput.stories'
 `ANumberInput` is a wrapper around the HTML `<input>` element with all the native functionality.
 Its props and styles are consistent with other inputs in our design system.
 
-<ArgTypes of={AInput} />
-
+<ArgTypes of={ANumberInput} />
 <ArgTypes of={NumberInputStories} />
 
 ## Labelling the input
